refactor(seasons): extract shared status config
Dedup season status colors and labels into lib/season-status-config.ts:
- STATUS_BADGE_STYLES for status-badge.tsx
- TIMELINE_DOT_COLORS for seasons list page
- LIVE_PILL_STYLE / FINAL_PILL_STYLE for season detail page
